Title: 바이러스 검사, Time: 1798ms, Memory: 158MB, Status: Passed - CodeTree

+++ 241227/바이러스 검사/virus-detector.java
@@ -1,4 +1,5 @@
 import java.util.*;
+import java.math.*;
 
 public class Main {
     public static void main(String[] args) {
@@ -18,7 +19,7 @@ public class Main {
         int head = scanner.nextInt();
         // 팀원이 검사 가능한 수
         int tail = scanner.nextInt();
-        Integer count = 0;
+        long count = 0;
 
         for(int i=0;i<restCount;i++){
             int temp = custCount[i];
